fix some problems, add output to UART3

+++ Task2/prj_files/inc/uTimer_init.h
@@ -3,5 +3,5 @@
 #include "stdint.h"
 #define TIMER				MDR_TIMER1
 #define TIMER_IRQ		TIMER1_IRQn
-void m_setUpTimer(uint8_t req_time, uint8_t req_CPUfreq);
+void m_setUpTimer(float req_time, uint8_t req_CPUfreq);
 #endif // __UTIMER_INIT
